Update: Connected and Create DataBase

+++ RestWithASPNETUdemy/Model/Services/Implementations/IPersonService.cs
@@ -5,8 +5,8 @@
         Person Create(Person person);
         Person Update(Person person);
         List<Person> FindAll();
-        Person FindById(long id);
-        void Delete(long id);
+        Person FindById(int id);
+        void Delete(int id);
 
     }
 }
